requirements.txt
praw==7.7.1
python-dotenv==1.0.1
openai==1.31.0
tqdm==4.66.4   
